feat: Alarm 도메인의 Service 클래스에서 AlarmResponseCode 적용

+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/alarm/service/AlarmService.java
@@ -15,10 +15,10 @@ import com.hertz.hertz_be.domain.alarm.repository.AlarmRepository;
 import com.hertz.hertz_be.domain.alarm.repository.UserAlarmRepository;
 import com.hertz.hertz_be.domain.channel.entity.SignalRoom;
 import com.hertz.hertz_be.domain.channel.entity.enums.MatchingStatus;
-import com.hertz.hertz_be.domain.channel.exception.UserNotFoundException;
 import com.hertz.hertz_be.domain.user.entity.User;
 import com.hertz.hertz_be.domain.user.repository.UserRepository;
-import com.hertz.hertz_be.global.exception.InternalServerErrorException;
+import com.hertz.hertz_be.global.common.NewResponseCode;
+import com.hertz.hertz_be.global.exception.BusinessException;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.data.domain.Page;
@@ -27,7 +27,6 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.support.TransactionSynchronization;
-import org.springframework.transaction.support.TransactionSynchronizationAdapter;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 import java.time.LocalDateTime;
@@ -52,7 +51,11 @@ public class AlarmService {
     @Transactional
     public void createNotifyAlarm(CreateNotifyAlarmRequestDto dto, Long userId) {
         User notifyWriter = userRepository.findById(userId)
-                .orElseThrow(UserNotFoundException::new);
+                .orElseThrow(() -> new BusinessException(
+                        NewResponseCode.NOT_FOUND.getCode(),
+                        NewResponseCode.NOT_FOUND.getHttpStatus(),
+                        "요청한 사용자를 찾을 수 없습니다."
+                ));
 
         AlarmNotification alarmNotification = AlarmNotification.builder()
                 .title(dto.getTitle())
@@ -181,7 +184,7 @@ public class AlarmService {
                                 report.getCreatedAt().toString()
                         );
                     } else {
-                        throw new InternalServerErrorException();
+                        throw new RuntimeException();
                     }
                 })
                 .collect(Collectors.toList());
@@ -205,8 +208,13 @@ public class AlarmService {
 
     @Transactional
     public void deleteAlarm(Long alarmId, Long userId) {
-        userRepository.findById(userId)
-                .orElseThrow(UserNotFoundException::new);
+        if (!userRepository.existsById(userId)) {
+            throw new BusinessException(
+                    NewResponseCode.NOT_FOUND.getCode(),
+                    NewResponseCode.NOT_FOUND.getHttpStatus(),
+                    "요청한 사용자를 찾을 수 없습니다."
+            );
+        }
 
         alarmRepository.deleteById(alarmId);
     }
